src/nolstudio/tests/com/gensym/nols/data/function/SubtractTest.java
package com.gensym.nols.data.function;

import static org.junit.Assert.assertEquals;

import java.util.Vector;

import org.junit.Before;
import org.junit.Test;

import com.gensym.math.MathException;

public class SubtractTest
{
	private Subtract testObj;

	@Before
	public void initObject()
	{
		testObj = new Subtract();
	}
	/**
	 * Method to test getName()
	 */
	@Test
	public void testGetName()
	{
		assertEquals("Subtract", testObj.getName());
	}

	/**
	 * Method to test getDescription()
	 */
	@Test
	public void testGetDescription()
	{
		final String description = "C = Subtract(A,B).  Returns C = A - B";
		assertEquals(description, testObj.getDescription());
	}

	/**
	 * Method to test getMinimumInputs
	 */
	@Test
	public void testGetMinimumInputs()
	{
		assertEquals(2, testObj.getMinimumInputs());
	}

	/**
	 * Method to test getMaximumInputs
	 */
	@Test
	public void testGetMaximumInputs()
	{
		assertEquals(2, testObj.getMaximumInputs());
	}

	/**
	 * Method to test getNumberOfOutputs
	 */
	@Test
	public void testGetNumberOfOutputs()
	{
		assertEquals(1, testObj.getNumberOfOutputs(12));
	}

	/**
	 * Method to test evaluate()
	 * @throws MathException
	 */
	@Test
	public void testEvaluate() throws MathException
	{
		final Vector<double[]> inputVector = new Vector<>();
		final double[] in1 = {10.0, 20.0, 40.0, 90.0};
		final double[] in2 = {6.0, 120.0, 20.0, -90.0};
		inputVector.addElement(in1);
		inputVector.addElement(in2);

		final double[] outputVals = new double[in1.length];
		final Vector<double[]> outputVector = new Vector<>();
		outputVector.addElement(outputVals);
		final long[] t = {5000l, 6000l, 7000l, 8000l};

		testObj.evaluate(t, inputVector, outputVector);
		assertEquals(4.0, outputVals[0], 0);
		assertEquals(-100.0, outputVals[1], 0);
		assertEquals(20.0, outputVals[2], 0);
		assertEquals(180.0, outputVals[3], 0);
	}
}
